requirement.txt
streamlit
Pillow
numpy
tensorflow
google-generativeai
python-dotenv
